モジュール本体の名前空間・ファイルパスを変更
